[CORE/PERSISTENCE/GUI] - Smart Modules dev

+++ gemini-gui/src/main/java/it/at7/gemini/core/SmartModuleManagerInitImpl.java
@@ -2,11 +2,13 @@ package it.at7.gemini.core;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import it.at7.gemini.core.persistence.PersistenceEntityFilterUtilityService;
 import it.at7.gemini.exceptions.GeminiException;
 import it.at7.gemini.exceptions.GeminiGenericException;
 import it.at7.gemini.schema.Entity;
 import it.at7.gemini.schema.SmartModuleEntity;
 import it.at7.gemini.schema.SmartModuleEntityRef;
+import it.at7.gemini.schema.SmartSchemaWrapper;
 import it.at7.gemini.schema.smart.SmartSchema;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,6 +21,7 @@ import java.io.InputStream;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Scanner;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,40 +32,49 @@ public class SmartModuleManagerInitImpl implements SmartModuleManagerInit {
     private final ApplicationContext applicationContext;
     private final EntityManager entityManager;
     private final SchemaManagerInit schemaManagerInit;
+    private final PersistenceEntityFilterUtilityService persistenceEntityFilterUtilityService;
 
     @Autowired
-    public SmartModuleManagerInitImpl(ApplicationContext applicationContext, EntityManager entityManager, SchemaManagerInit schemaManagerInit) {
+    public SmartModuleManagerInitImpl(ApplicationContext applicationContext,
+                                      EntityManager entityManager,
+                                      SchemaManagerInit schemaManagerInit,
+                                      PersistenceEntityFilterUtilityService persistenceEntityFilterUtilityService) {
         this.applicationContext = applicationContext;
         this.entityManager = entityManager;
         this.schemaManagerInit = schemaManagerInit;
+        this.persistenceEntityFilterUtilityService = persistenceEntityFilterUtilityService;
     }
 
     @Override
-    public void initializeSmartModulesRecords(LinkedHashMap<SmartModule, SmartSchema> schemasByModule, Transaction transaction) throws GeminiException {
+    public void initializeSmartModulesRecords(LinkedHashMap<SmartModule, SmartSchemaWrapper> schemasByModule, Transaction transaction) throws GeminiException {
         EntityOperationContext operationContext = schemaManagerInit.getOperationContextForInitSchema();
-        for (SmartModule smartModule : schemasByModule.keySet()) {
-            SmartModuleEntity mEntity = SmartModuleEntity.of(smartModule);
+        for (Map.Entry<SmartModule, SmartSchemaWrapper> entry : schemasByModule.entrySet()) {
+            SmartModule module = entry.getKey();
+            SmartSchemaWrapper schemaWrapper = entry.getValue();
+
+            SmartModuleEntity mEntity = SmartModuleEntity.of(module, schemaWrapper.getSmartSchemaAsString());
             EntityRecord er = mEntity.toEntityRecord();
             this.entityManager.putOrUpdate(er, operationContext, transaction);
         }
-        Entity smartModuleEntity = entityManager.getEntity(SmartModuleEntityRef.NAME);
-        List<String> smarNames = schemasByModule.keySet().stream().map(SmartModule::getName).map(String::toUpperCase).collect(Collectors.toList());
-        // TODO persistence interface to have a not IN filter query
-        String condition = " name NOT IN (:smart_names)";
-        Map<String, Object> params = Map.of("smart_names", smarNames);
-        List<EntityRecord> records = this.entityManager.getRecordsMatching(smartModuleEntity, FilterContext.withPersistenceQueryParam(condition, params), operationContext, transaction);
-        for (EntityRecord record : records) {
-            entityManager.delete(record, operationContext, transaction);
-        }
+        if (!schemasByModule.isEmpty()) {
+            Entity smartModuleEntity = entityManager.getEntity(SmartModuleEntityRef.NAME);
+            List<String> smartNames = schemasByModule.keySet().stream().map(SmartModule::getName).map(String::toUpperCase).collect(Collectors.toList());
+            String condition = persistenceEntityFilterUtilityService.notContainsCondition("name", "smart_names");
+            Map<String, Object> params = Map.of("smart_names", smartNames);
+            List<EntityRecord> records = this.entityManager.getRecordsMatching(smartModuleEntity, FilterContext.withPersistenceQueryParam(condition, params), operationContext, transaction);
+            for (EntityRecord record : records) {
+                entityManager.delete(record, operationContext, transaction);
+            }
 
-        for (Map.Entry<SmartModule, SmartSchema> moduleSchemaPair : schemasByModule.entrySet()) {
-            handleGUIEntityRecords(moduleSchemaPair, transaction);
+            for (Map.Entry<SmartModule, SmartSchemaWrapper> moduleSchemaPair : schemasByModule.entrySet()) {
+                handleGUIEntityRecords(moduleSchemaPair, transaction);
+            }
         }
     }
 
     @Override
-    public LinkedHashMap<SmartModule, SmartSchema> loadSmartResources(List<SmartModule> modulesInOrder) throws GeminiException {
-        LinkedHashMap<SmartModule, SmartSchema> smartSchema = new LinkedHashMap<>();
+    public LinkedHashMap<SmartModule, SmartSchemaWrapper> loadSmartResources(List<SmartModule> modulesInOrder) throws GeminiException {
+        LinkedHashMap<SmartModule, SmartSchemaWrapper> smartSchemas = new LinkedHashMap<>();
         try {
             for (SmartModule module : modulesInOrder) {
                 if (!module.isDynamic()) {
@@ -73,21 +85,26 @@ public class SmartModuleManagerInitImpl implements SmartModuleManagerInit {
                         logger.info("Smart Schema definition found for Smart Module {}: location {}", module.getName(), location);
 
                         ObjectMapper yamlMapper = new ObjectMapper(new YAMLFactory());
-                        smartSchema.put(module, yamlMapper.readValue(schemaStream, SmartSchema.class));
+                        SmartSchema smartSchema = yamlMapper.readValue(schemaStream, SmartSchema.class);
+
+                        Scanner scanner = new Scanner(resource.getInputStream()).useDelimiter("\\A");
+                        String smartSchemaString = scanner.hasNext() ? scanner.next() : "";
+
+                        smartSchemas.put(module, SmartSchemaWrapper.of(smartSchema, smartSchemaString));
                     } else {
                         logger.info("NO Smart Schema definition found for Smart Module {}: location {}", module.getName(), location);
                     }
                 }
             }
-            return smartSchema;
+            return smartSchemas;
         } catch (Exception e) {
             throw GeminiGenericException.wrap(e);
         }
     }
 
-    private void handleGUIEntityRecords(Map.Entry<SmartModule, SmartSchema> moduleSchemaEntry, Transaction transaction) throws GeminiException {
+    private void handleGUIEntityRecords(Map.Entry<SmartModule, SmartSchemaWrapper> moduleSchemaEntry, Transaction transaction) throws GeminiException {
         SmartModule smartModule = moduleSchemaEntry.getKey();
-        SmartSchema smartSchema = moduleSchemaEntry.getValue();
+        SmartSchema smartSchema = moduleSchemaEntry.getValue().getSmartSchema();
         addALLRecords(smartSchema.getEntityGUIRecords(smartModule), transaction);
         addALLRecords(smartSchema.getFieldGUIRecords(smartModule), transaction);
     }
